fix: give the head-link tests a class that gets a database
MediaWiki decides whether a test class needs a database once, for the whole
class, so a group on the method changed nothing and the three that write and
read a licenses copy still ran with the backend disabled. They move to a class
of their own, grouped, which also keeps the twenty tests in MainTest that need
no database from paying for one.
